Add files via upload
Here we used K-means analysis and K-mean cluster plot to visualize the expression of gene groups categorized by their k-means. K-mean analysis was used to to determine the points, centroids, and distance between centroids. To visualize the results, a cluster plot was created that showed differential expression between gene groups when treated with IL-6, IL-1, or both.